Workaround issue with bad schema validation w/ libxml2 v2.7.8
A bug was encountered where parsing of durations would raise an exception in
lxml.  The bug was tracked down to the depths of libxml2 where the duration
string was parsed.

The issue has been fixed in the git repository (2011-11-10), but v2.7.9 has
not yet been released.

For more information see the following links:

- https://bugzilla.gnome.org/show_bug.cgi?id=653724
- http://git.gnome.org/browse/libxml2/commit/?id=bbcf1275fec28915b170db31fa9ea9bed0ea39a4

To use the workaround add the following before using soapbox:

    from soapboax import xsd
    xsd.VALIDATE_ON_PARSE = False