use "lists" for service methods (this reverts 6be29fb9464367f455a86e2b905a72d61e8b2c0b)
Even though the internal soapbox API is slightly nicer if methods are
stored in a dict. However this approach has some drawbacks:
- soapbox users need to duplicate some information (xsd.Method has
  an attribute "operationName" already)
- as a result the service definition is more verbose

The soapbox user API is IMHO more important than the internal code so
I'm reverting this commit. After a few emails it seems that Thomas
is not opposing that change completely so I think we're good to go :-)