add minimal stub code to support custom choice elements
The current model code does not allow representing XSD types like:
<xs:complexType>
	<xs:choice maxOccurs="5000">
		<xs:element ref="tns:foo"/>
		<xs:element ref="tns:bar"/>
	</xs:choice>
</xs:complexType>

With the changes in this commit that becomes possible (using some custom
types and a bit of hackery). I'm not sure yet how this should work in a
final version but until I need to support it in my custom code.