requests: substitute for httplib2
In the past there were issues with CA certificates and socket timeouts
that affected httplib2, particularly if certificates on a system
provided in the package manager were not updated and stale. There have
also been issues such as support for SNI.

Since soapbox/soapfish was originally developed, the requests package
became available and simplified things as well as smoothing out a number
of the issues that existed. It has somewhat become the de facto package
to use when making web requests.

License: BSD