wsdl2py: fix detection of xml schema namespaces
In some cases we have encountered a different namespace prefix in use
for schemas with xsd: used on the root element and xs: used on the
embedded schema elements, i.e.

    <wsdl:definitions xmlns:xsd="http://www.w3.org/2001/XMLSchema" ...>
      <wsdl:types>
        <xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema" ...>
        </xs:schema>
      </wsdl:types>
    </wsdl:definitions>

We recursively look for any schema elements and update the namespace map
before searching for namespace prefixes.

License: BSD