wsdl: remove remaining reverse reference code.
Refactor the lookup of message objects and headers so that we can remove
the complicated reverse reference approach used in the WSDL classes.

This removes one of the barriers to recombining the classes for the
definitions of WSDL over SOAP 1.1 and 1.2 and also makes the resulting
code much less complicated.

License: BSD