xsd: initial ordering support for complex types.
We should implement support for ordering and hashing of all types.
Initially we have implemented limited support for complex types which
was required to check for equality of schemas. Note that this
implementation is currently very naive and converts back to the raw XML
from the element tree where it can. Naturally this is rather inefficient.

License: BSD