tree-wide: improve entry points for easier testing.
Make it possible to pass command line arguments to main() functions to
make writing tests easier. Also tidied up handling of command line
argument parsing and made it possible to specify output files or use
stdin to pass in the input document.

Because nose replaces stdout by default with StringIO() the buffer
attribute is unavailable for binary output which causes some tests
to fail. This simplifies the check for the buffer object and falls back
to the vanilla sys.stdout which now seems to work for all our cases in
Python 2 and 3, as well as when running nosetests without --nocapture.

License: BSD