misc: rename known_files to known_paths
This used to be tracked as known_namespaces until 9e9cb6b but much work
has been done that expects that all WSDL and XSD are local files. We
need to restore support for WSDL and XSD hosted remotely. This is an
initial step to make reference to 'paths' rather than 'files'.

License: BSD