soap: refactor location rewriting in stub (#91)
This eliminates the `base_url` keyword argument in favour of `location`
accepting a callable to allow for more advanced re-writing of the
location. One example is to replace parts of the path based on the host
that the request will be sent to. The callable passed to the location
must accept two arguments - the template and context containing default
parameters for scheme and host.