dispatcher: pass request preparation through middleware stack.
We may want to be able to log errors that occur during preparation of
the SOAP request object, e.g. syntax errors, but this requires passage
through the middleware stack to be able to use ExceptionLogger with
SOAPError.

Exceptions are not converted to SOAP faults when using the parsed body
unless we use the ExceptionToSoapFault middleware, but exceptions when
parsing the envelope are converted. We should consider whether to
standardise this behaviour to convert all or no errors to SOAP faults,
whether to make the exceptions to convert customisable, and whether to
enable the ExceptionToSoapFault middleware by default. If all of this is
done, we should then probably allow the original errors encountered
during parsing, e.g. DocumentInvalid, XMLSyntaxError, to be raised and
converted by the middleware without special handling.